+++ Platforms/Arguments/Package.swift
@@ -25,7 +25,6 @@ let package = Package(
   ],
   dependencies: [
     .package(name: "swift-tui", path: "../.."),
-    .package(name: "SwiftTUICLI", path: "../CLI"),
     .package(url: "https://github.com/apple/swift-argument-parser.git", from: "1.5.0"),
   ],
   targets: [
@@ -33,7 +32,6 @@ let package = Package(
       name: "SwiftTUIArguments",
       dependencies: [
         .product(name: "SwiftTUI", package: "swift-tui"),
-        .product(name: "SwiftTUICLI", package: "SwiftTUICLI"),
         .product(name: "ArgumentParser", package: "swift-argument-parser"),
       ],
       swiftSettings: swiftSettings()