+++ Platforms/WebHost/Sources/SwiftTUIWebHostCLI/SwiftTUIWebHostCLI.swift
@@ -1,2 +1,3 @@
 @_exported import SwiftTUI
+@_exported import SwiftTUIArguments
 @_exported import SwiftTUIWebHost